% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{addClonotype}
\alias{addClonotype}
\title{Combine Clonotype Object}
\usage{
addClonotype(clonotype, clonotype2)
}
\arguments{
\item{consensus}{class. an object of the clonotype class}

\item{consensus2}{class. an object of the clonotype class}
}
\value{
a combined clonotype class
}
\description{
Combine Clonotype Object
}
\examples{

}
