% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{addConsensus}
\alias{addConsensus}
\title{Combine Consensus Object}
\usage{
addConsensus(consensus, consensus2)
}
\arguments{
\item{consensus}{class. an object of the consensus class}

\item{consensus2}{class. an object of the consensus class}
}
\value{
a combined consensus class
}
\description{
Combine Consensus Object
}
\examples{

}
