% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.r
\name{appendFile}
\alias{appendFile}
\title{Append Large Files}
\usage{
appendFile(
  base,
  append,
  base_sep = "\\t",
  base_col = 1,
  base_by_line = FALSE,
  append_sep = "\\t",
  append_by_line = FALSE,
  append_col = 1,
  append_header = TRUE,
  output = "result.txt",
  output_sep = "\\t",
  fill = "-"
)
}
\arguments{
\item{base}{character.}

\item{append}{character.}

\item{base_sep}{character.}

\item{base_col}{numeric.}

\item{base_by_line}{logical.}

\item{append_sep}{character.}

\item{append_by_line}{logical.}

\item{append_col}{numeric.}

\item{append_header}{logical.}

\item{output}{character.}

\item{output_sep}{character.}

\item{fill}{character.}
}
\value{

}
\description{
Append Large Files
}
\examples{
print('waiting')

}
