% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refIMGT.r
\name{build_IMGT_reference}
\alias{build_IMGT_reference}
\title{Build IMGT database reference}
\usage{
build_IMGT_reference(outdir = NULL, verbose = TRUE)
}
\arguments{
\item{outdir}{character. Default \code{getwd()}}

\item{verbose}{logical. Default TRUE}
}
\value{
if success, return TRUE
}
\description{
Download reference sequences from the IMGT (the international ImMunoGeneTics information system,
http://www.imgt.org) database and split the sequences by species.
}
\examples{
\donttest{build_IMGT_reference('IMGT_reference', verbose = FALSE)}

}
