% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReference.r
\name{build_viruSite_reference}
\alias{build_viruSite_reference}
\title{Build viruSite database reference}
\usage{
build_viruSite_reference(
  outdir = NULL,
  method = NULL,
  ftp_method = NULL,
  verbose = TRUE
)
}
\arguments{
\item{outdir}{character. Default \code{getwd()}}

\item{method}{character. Method to be used for downloading html files, equal to \code{download.file}. Default 'libcurl'}

\item{ftp_method}{character. Method to be used for downloading ftp files (.fa and .gtf), equal to \code{download.file}. Default 'curl'}

\item{verbose}{logical. Default \code{TRUE}}
}
\value{
if success, return \code{TRUE}
}
\description{
Download reference sequences from the viruSite database and make genome and gene gtf.
}
\examples{
\donttest{build_viruSite_reference('viruSite')}

}
