% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cbinds}
\alias{cbinds}
\title{Combine Two Data-frame by Columns}
\usage{
cbinds(F1, F2, fill = 0)
}
\arguments{
\item{F1}{data.frame.}

\item{F2}{data.frame.}

\item{fill}{character/numeric. Default 0}
}
\value{
a combined data.frame
}
\description{
Combine two data.frame by columns by filling in missing rows from each other based on \code{rownames}.
}
\examples{
F1 = data.frame(A = seq(10), B = seq(10), row.names = seq(10))
F2 = data.frame(C = seq(5),  D = seq(5),  row.names = 3:7)
cbinds(F1, F2)

}
