% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkDup}
\alias{checkDup}
\title{Check Duplicated Value}
\usage{
checkDup(x)
}
\arguments{
\item{x}{vector.}
}
\value{
a logical vector indicating which elements are duplicated.
}
\description{
Check Duplicated Value
}
\examples{
checkDup(c(1:5, 2:10))

}
