% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkIn}
\alias{checkIn}
\title{Check a Name in Names}
\usage{
checkIn(obs, pool)
}
\arguments{
\item{obs}{vector.}

\item{pool}{vector.}
}
\value{
a inner name or \code{NULL}
}
\description{
Check a Name in Names
}
\examples{
checkIn(2, 1:3)
checkIn(4, 1:3)

}
