% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{checkName}
\alias{checkName}
\title{Check Name in an VDJ Object}
\usage{
checkName(vdj, name = NULL)
}
\arguments{
\item{vdj}{an object of vdj.}

\item{name}{character.}
}
\value{

}
\description{
Check Name in an VDJ Object
}
\examples{

}
