% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{checkSub}
\alias{checkSub}
\title{Check Subset Value}
\usage{
checkSub(x, i)
}
\arguments{
\item{x}{vector.}

\item{i}{index.}
}
\value{
a subset value
}
\description{
Check Subset Value
}
\examples{
checkSub(NULL, 1)

}
