% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Circos}
\alias{Circos}
\title{Plot Circos}
\usage{
Circos(VJpair, col_grid = NULL, col_link = NULL)
}
\arguments{
\item{VJpair}{character.}

\item{col_grid}{character.}

\item{col_link}{character.}
}
\value{

}
\description{
Plot Circos
}
\examples{
VJpair = c('TRAV12-1~TRAJ35', 'TRAV12-1~TRAJ35',  'TRAV19~TRAJ35', 
           'TRAV41~TRAJ35',   'TRAV36DV7~TRAJ21', 'TRAV8-3~TRAJ35')
Circos(VJpair)

}
