% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{cloneAbundance}
\alias{cloneAbundance}
\title{Estimate clonotype abundance}
\usage{
cloneAbundance(vdj, names = NULL, plot = TRUE, save = TRUE, out.pref = NULL)
}
\arguments{
\item{vdj}{an object of VDJ.}

\item{names}{character.}

\item{plot}{logical.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Estimate clonotype abundance
}
\examples{

}
