% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{cloneShanno}
\alias{cloneShanno}
\title{Estimate Clonotype Shannon}
\usage{
cloneShanno(
  vdj,
  names = NULL,
  plot = TRUE,
  colors = NULL,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{an object of vdj}

\item{names}{character}

\item{plot}{logical}

\item{save}{logical}

\item{out.pref}{character}
}
\value{

}
\description{
Estimate Clonotype Shannon
}
\examples{

}
