% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{CloneSize}
\alias{CloneSize}
\title{Estimate Clonotype clone size}
\usage{
CloneSize(
  vdj,
  names = NULL,
  sep = 1:3,
  plot = TRUE,
  colors = NULL,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{an object of VDJ.}

\item{names}{character. sample or group names.}

\item{sep}{integer.}

\item{plot}{logical.}

\item{colors}{character.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{
results of clone size
}
\description{
Estimate Clonotype clone size
}
\examples{
VDJ = readRDS('VDJ.rds')
CloneSize(VDJ)

}
