% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{CloneSizeFisher}
\alias{CloneSizeFisher}
\title{Fisher Test for Clone Size}
\usage{
CloneSizeFisher(
  cloneSize,
  control = NULL,
  treatment = NULL,
  types = NULL,
  pval = 0.05,
  plot = TRUE,
  save = TRUE
)
}
\arguments{
\item{cloneSize}{data.frame.}

\item{control}{character.}

\item{treatment}{character.}

\item{types}{character.}

\item{pval}{numeric.}

\item{plot}{logical.}

\item{save}{logical.}
}
\value{
a fisher test data.frame
}
\description{
Fisher Test for Clone Size
}
\examples{

}
