% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{clonotype-class}
\alias{clonotype-class}
\alias{clonotype}
\title{The Clonotype Class}
\value{
An object of the clonotype class
}
\description{
The Clonotype Class
}
\section{Slots}{

\describe{
\item{\code{ID}}{character.}

\item{\code{ConsenID}}{character.}

\item{\code{Cells}}{numeric.}

\item{\code{Barcodes}}{character.}

\item{\code{CDR3dna}}{character.}

\item{\code{CDR3aa}}{character.}
}}

