% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{clonotypeAbundance}
\alias{clonotypeAbundance}
\title{Estimate clonotype abundance}
\usage{
clonotypeAbundance(vdj, names = NULL, plot = TRUE, save = TRUE)
}
\arguments{
\item{vdj}{an object of VDJ.}

\item{names}{character.}

\item{plot}{logical.}

\item{save}{logical.}
}
\value{

}
\description{
Estimate clonotype abundance
}
\examples{

}
