% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{clonotypeShanno}
\alias{clonotypeShanno}
\title{Estimate Clonotype Shannon}
\usage{
clonotypeShanno(vdj, names = NULL, plot = TRUE, save = TRUE)
}
\arguments{
\item{vdj}{}

\item{names}{}

\item{plot}{}

\item{save}{}
}
\value{

}
\description{
Estimate Clonotype Shannon
}
\examples{

}
