% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{consensus-class}
\alias{consensus-class}
\alias{consensus}
\title{The consensus class}
\value{
An object of the consensus class
}
\description{
The consensus class
}
\section{Slots}{

\describe{
\item{\code{Vgene}}{character.     V gene, eg.: TRAV1*01}

\item{\code{Dgene}}{character.     D gene, eg.: TRAD1*01}

\item{\code{Jgene}}{character.     J gene, eg.: TRAJ1*01}

\item{\code{Cgene}}{character.     C gene, eg.: TRAC1*01}

\item{\code{CDR1dna}}{character.     CDR1 nucleic acid sequence, eg.: TCTGAACACAACCGC}

\item{\code{CDR2dna}}{character.     CDR2 nucleic acid sequence, eg.: TTCCAGAATGAAGCTCAA}

\item{\code{CDR3dna}}{character.     CDR3 nucleic acid sequence, eg.: TGTGCCAGCAGCCTACGCAACGAGCAGTACTTC}

\item{\code{CDR3aa}}{character.     CDR3 amino acid sequence, eg.: CASSPTPGEATDTQYF}

\item{\code{Count}}{numeric.       Read counts, eg.: 200}

\item{\code{ID}}{character.     Consensus id, eg.: Sample1_consensus1}

\item{\code{CDR3germlineSimilarity}}{numeric. CDR3 germline similarity score, eg.: 80}

\item{\code{FullLength}}{logical.    Whether the vdj gene is complete, eg.: TRUE}
}}

