% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{ConsensusFromDataframe}
\alias{ConsensusFromDataframe}
\title{Create Consensus Object from Data Frame}
\usage{
ConsensusFromDataframe(Df, properties = NULL, verbose = TRUE)
}
\arguments{
\item{Df}{data.frame.}

\item{properties}{list.}

\item{verbose}{logical.}
}
\value{
an object of consensus
}
\description{
Create Consensus Object from Data Frame
}
\examples{
# file paths
airr_file   = system.file('extdata', '10x_airr_rearrangement.tsv.gz', package = 'TrustVDJ')
contig_file = system.file('extdata', '10x_all_contig_annotations.csv.gz', package = 'TrustVDJ')
# read file
data        = Read10x(airr_file = airr_file, contig_file = contig_file)
# create consensus object
consensus   = ConsensusFromDataframe(data, properties = list(
 ConsenID = 'raw_consensus_id', ClonoID = 'clone_id', Barcode = 'cell_id', 
 Vgene = 'v_call', Dgene = 'd_call', Jgene = 'j_call', Cgene = 'c_call',
 UMIs = 'duplicate_count', Reads = 'consensus_count') )
consensus

}
