% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{CreateVdjObject}
\alias{CreateVdjObject}
\title{Create a VDJ Object from a VDJ sample object list}
\usage{
CreateVdjObject(samples, group = NULL, project = NULL, verbose = TRUE)
}
\arguments{
\item{samples}{list.}

\item{group}{list.}

\item{project}{character.}

\item{verbose}{logical.}
}
\value{
An object of VDJ
}
\description{
Create a VDJ Object from a VDJ sample object list
}
\examples{

}
