% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{CreateVdjSample}
\alias{CreateVdjSample}
\title{Create a VDJ Sample Object from Data Frame}
\usage{
CreateVdjSample(
  Df,
  properties = NULL,
  name = NULL,
  rm_allele = FALSE,
  unique_clonotype = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Df}{data.frame.}

\item{properties}{list.}

\item{name}{character.}

\item{rm_allele}{logical.}

\item{unique_clonotype}{logical.}

\item{verbose}{logical.}
}
\value{
An object of VDJ sample
}
\description{
Create a VDJ Sample Object from Data Frame
}
