% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{df_chain}
\alias{df_chain}
\title{data.frame a single chain information}
\usage{
df_chain(chain)
}
\arguments{
\item{chain}{list. trust4 single chain information in a list}
}
\value{
a data.frame named by \code{chainName}
}
\description{
data.frame a single chain information
}
\examples{
df_chain(list('V', 'D', 'J', 'C', 'CDR3nt', 'CDR3aa', '60', 'id1', '98', '1'))

}
