% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.r
\name{.Read10x_clonotype}
\alias{.Read10x_clonotype}
\title{Read 10x clonotype file}
\usage{
.Read10x_clonotype(clonotype_file, verbose = TRUE)
}
\arguments{
\item{clonotype_file}{character. Path to clonotype file generated by 10x cellranger.}

\item{verbose}{logical. Print progress. Default is TRUE}
}
\value{
A VDJ information data.frame
}
\description{
Read 10x clonotype file
}
\examples{
clonotype_file = system.file('extdata', '10x_clonotypes.csv.gz', package = 'TrustVDJ')
clonotype = .Read10x_clonotype(clonotype_file = clonotype_file, verbose = FALSE)
head(clonotype)

}
