% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.r
\name{.Read10x_consensus}
\alias{.Read10x_consensus}
\title{Read 10x consensus file}
\usage{
.Read10x_consensus(consensus_file, verbose = TRUE)
}
\arguments{
\item{consensus_file}{character. Path to consensus_annotations file generated by 10x cellranger.}

\item{verbose}{logical. Print progress. Default is TRUE}
}
\value{
A VDJ information data.frame
}
\description{
Read 10x consensus file
}
\examples{
consensus_file = system.file('extdata', '10x_consensus_annotations.csv.gz', package = 'TrustVDJ')
consensus = .Read10x_consensus(consensus_file = consensus_file, verbose = FALSE)
head(consensus)

}
