% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.r
\name{.Read10x_contig}
\alias{.Read10x_contig}
\title{Read 10x contig file}
\usage{
.Read10x_contig(contig_file, verbose = TRUE)
}
\arguments{
\item{contig_file}{character. Path to contig_annotations file generated by 10x cellranger.}

\item{verbose}{logical. Print progress. Default is TRUE}
}
\value{
A VDJ information data.frame
}
\description{
Read 10x contig file
}
\examples{
contig_file = system.file('extdata', '10x_all_contig_annotations.csv.gz', package = 'TrustVDJ')
contig = .Read10x_contig(contig_file = contig_file, verbose = FALSE)
head(contig)

}
