% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTrust.r
\name{.ReadTrust_Report}
\alias{.ReadTrust_Report}
\title{Read TRUST4 report file.}
\usage{
.ReadTrust_Report(report_file = NULL, verbose = TRUE)
}
\arguments{
\item{report_file}{character. Path to report file generated by TRUST4.}

\item{verbose}{logical. Print progress. Default is TRUE}
}
\value{
A VDJ information data.frame
}
\description{
\code{.ReadTrust_Report} reads a report file generated by TRUST4.
Note that it could be xx_report.tsv but not xx_barcode_report.tsv.
(.gz supported)
}
\examples{
report_file = system.file('extdata', 'TRUST4_report.tsv.gz', package = 'TrustVDJ')
report = .ReadTrust_Report(report_file = report_file, verbose = FALSE)
head(report)

}
