% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{Download}
\alias{Download}
\title{Download files retryably}
\usage{
Download(x, name = NULL, sleep = NULL, outdir = NULL)
}
\arguments{
\item{x}{character. URLs.}

\item{name}{character. file names.}

\item{sleep}{numeric. retry interval.}

\item{outdir}{character. output directory.}
}
\value{
downloaded files
}
\description{
Download files retryably
}
\examples{
URLs = paste0('http://www.imgt.org//download/', c('V-QUEST/IMGT_V-QUEST_reference_directory',
         'GENE-DB/IMGTGENEDB-ReferenceSequences.fasta-nt-WithGaps-F+ORF+inframeP'))
# Download(URLs, c('vdj_species.html', 'IMGT_download.fa') )

}
