% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{fetchCdr3}
\alias{fetchCdr3}
\title{Fetch CDR3dna or CDR3aa in a Consensus Class}
\usage{
fetchCdr3(consensus, type = "CDR3dna")
}
\arguments{
\item{consensus}{class.     an object of the consensus class}

\item{type}{character. CDR3dna or CDR3aa}
}
\value{

}
\description{
Fetch CDR3dna or CDR3aa in a Consensus Class
}
\examples{
CDR3 = fetchCdr3(consensus, 'CDR3aa')
head(CDR3)

}
