% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{fetchVDJ}
\alias{fetchVDJ}
\title{Fetch VDJ Genes in a Consensus Class}
\usage{
fetchVDJ(consensus, type = c("V", "D", "J", "C"))
}
\arguments{
\item{consensus}{class. an object of the consensus class}

\item{type}{character. Among \code{V, D, J, C}}
}
\value{

}
\description{
Fetch VDJ Genes in a Consensus Class
}
\examples{
gene = fetchVDJ(consensus)
head(gene)

}
