% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{fetchVJab}
\alias{fetchVJab}
\title{Fetch VJ-AB in a Consensus Class}
\usage{
fetchVJab(consensus, clonotype)
}
\arguments{
\item{consensus}{class. an object of the consensus class}
}
\value{

}
\description{
Fetch VJ-AB in a Consensus Class
}
\examples{
vjAB = fetchVJab(consensus, clonotype)
head(vjAB)

}
