% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{fetchVJpair}
\alias{fetchVJpair}
\title{Fetch VJ Pair in a Consensus Class}
\usage{
fetchVJpair(consensus)
}
\arguments{
\item{consensus}{class. an object of the consensus class}
}
\value{

}
\description{
Fetch VJ Pair in a Consensus Class
}
\examples{
vj = fetchVJpair(consensus)
head(vj)

}
