% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{filterGtf}
\alias{filterGtf}
\title{Filter gtf file by target chromosomes}
\usage{
filterGtf(gtf, chromosomes = NULL, out = NULL, sep = "\\t", verbose = TRUE)
}
\arguments{
\item{gtf}{character. Path to gtf file.}

\item{chromosomes}{character. Target chromosomes.}

\item{out}{character. Output file.}

\item{sep}{character. Each gtf line separator. Default \code{'\t'}}

\item{verbose}{logical. Print progress. Default TRUE}
}
\value{
if success, return \code{TRUE}
}
\description{
\code{filterGtf} reads gtf file and filter it by target chromosomes.
(.gz supported)
}
\examples{
gtf = system.file('extdata', 'IMGT_Homo_sapiens.gtf.gz', package = 'TrustVDJ')
filterGtf(gtf, c('IGHD1-1*01', 'TRDD1*01'), 'filter.gtf')

}
