% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{findCDR3Motif}
\alias{findCDR3Motif}
\title{Find CDR3 Motif in K-mer and Position}
\usage{
findCDR3Motif(
  cdr3,
  kmers = NULL,
  position = NULL,
  min.freq = 3,
  verbose = TRUE
)
}
\arguments{
\item{cdr3}{character.}

\item{kmers}{numeric.}

\item{position}{numeric.}

\item{min.freq}{numeric.}

\item{verbose}{logical.}
}
\value{

}
\description{
Find CDR3 Motif in K-mer and Position
}
\examples{
CDR3  = TableCDR3(vdj, type = 'CDR3aa')
CDR3  = rep(CDR3$CDR3, CDR3$TotalCell)
Motif = findCDR3Motif(CDR3, 3, c(4, -3))
head(Motif) 

}
