% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{findListName}
\alias{findListName}
\title{Find List Name}
\usage{
findListName(x, list)
}
\arguments{
\item{x}{character.}

\item{list}{list.}
}
\value{
list name
}
\description{
Find List Name
}
\examples{
list = list(N1 = c('a', 'b'), N2 = c('c', 'd'))
findListName(c('a', 'c'), list)

}
