% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{findMotif}
\alias{findMotif}
\title{Find Characters Motif in K-mer}
\usage{
findMotif(x, kmer = 3)
}
\arguments{
\item{x}{character.}

\item{kmer}{numeric.}
}
\value{

}
\description{
Find Characters Motif in K-mer
}
\examples{
findMotif(c('abcdefg', 'abcdabcd'))

}
