% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{Fisher}
\alias{Fisher}
\title{Fisher's Exact Test}
\usage{
Fisher(treatment, control, base = 2)
}
\arguments{
\item{treatment}{numeric.}

\item{control}{numeric.}

\item{base}{numeric. Default \code{2}}
}
\value{
a data.frame of Fisher's exact test results.
}
\description{
Fisher's Exact Test
}
\examples{
treatment = c(10:1)
control   = c(1:10)
Fisher(treatment, control)

}
