% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{GroupCDR3byMotif}
\alias{GroupCDR3byMotif}
\title{Group CDR3 by motif}
\usage{
GroupCDR3byMotif(CDR3, motif)
}
\arguments{
\item{CDR3}{character.}

\item{motif}{character.}
}
\value{

}
\description{
Group CDR3 by motif
}
\examples{
CDR3  = c('CSALSWDKPEETQYF', 'CSALSADKPEETQYF', 
          'CSALSWDAPEETQYF', 'CSALSWDKPEATQYF')
motif = c('EET', 'DKP')
GroupCDR3byMotif(CDR3, motif)

}
