% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{htmlText}
\alias{htmlText}
\title{Get HTML text content}
\usage{
htmlText(file, css = "body")
}
\arguments{
\item{file}{character. Path to HTML file.}

\item{css}{character. HTML elements (one of css) to select. Default \code{'body'}}
}
\value{
A character vector.
}
\description{
Get HTML text content
}
\examples{
\donttest{
Download('http://www.virusite.org', 'test.html')
content = htmlText('test.html')
head(content)
}

}
