% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Logo}
\alias{Logo}
\title{Plot character logo}
\usage{
Logo(x, col = NULL, out = NULL)
}
\arguments{
\item{x}{character.}

\item{out}{character.}
}
\value{

}
\description{
Plot character logo
}
\examples{
Logo(c('hahaha', 'hehehe'))

}
