% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makePct}
\alias{makePct}
\title{Show Number as Percentage}
\usage{
makePct(x, digits = 3)
}
\arguments{
\item{x}{vector.  a vector of number}

\item{digits}{numeric. default \code{3}}
}
\value{
percentage
}
\description{
Show Number as Percentage
}
\examples{
makePct(.253)

}
