% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{MergeClonotype}
\alias{MergeClonotype}
\title{Merge Clonotype of VDJ Sample Objects}
\usage{
MergeClonotype(samples, verbose = TRUE)
}
\arguments{
\item{samples}{list.}

\item{verbose}{logical.}
}
\value{
an object of clonotype
}
\description{
Merge Clonotype of VDJ Sample Objects
}
\examples{

}
