% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{MergeConsensus}
\alias{MergeConsensus}
\title{Merge Consensus of VDJ Sample Objects}
\usage{
MergeConsensus(samples, verbose = TRUE)
}
\arguments{
\item{samples}{list.}

\item{verbose}{logical.}
}
\value{
an object of consensus
}
\description{
Merge Consensus of VDJ Sample Objects
}
\examples{

}
