% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{min.max}
\alias{min.max}
\title{Dispersion normalization}
\usage{
\method{min}{max}(x)
}
\arguments{
\item{x}{numeric.}
}
\value{
data in \link{0,1} after linear transformation
}
\description{
Dispersion normalization
}
\examples{
min.max(1:5)

}
