% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{none}
\alias{none}
\title{Check NA Values in a Data Frame}
\usage{
none(df)
}
\arguments{
\item{df}{data.frame}
}
\value{

}
\description{
Check NA Values in a Data Frame
}
\examples{
none(data.frame(c(1:3, 'None', 5), c(1, NA, '*', 4:5)))

}
