% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{objSize}
\alias{objSize}
\title{Report the Space Allocated for an Object in Mb}
\usage{
objSize(obj)
}
\arguments{
\item{obj}{an R object}
}
\value{
an estimate of the memory allocation attributable to the object in Mb.
}
\description{
Report the Space Allocated for an Object in Mb
}
\examples{
obj = 1:1024^2
objSize(obj)

}
