% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pick}
\alias{pick}
\title{Pick Field}
\usage{
pick(x, f = 1, exct = "\\\\|")
}
\arguments{
\item{x}{character.}

\item{f}{numeric. target field. Default \code{1}}

\item{exct}{pattern. Default \code{'\\|'}}
}
\value{
A character vector.
}
\description{
Pick Field
}
\examples{
char = c('Hello, miku!', 'Love? Love!')
pick(char, 2, ' ')

}
