% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Pie}
\alias{Pie}
\title{Plot Pie}
\usage{
Pie(x, y, fill = c("white", "#f8766d"), color = NULL)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{fill}{color}

\item{color}{color}
}
\value{

}
\description{
Plot Pie
}
\examples{
Pie(1:10, 5:12)

}
