% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rbinds}
\alias{rbinds}
\title{Combine Two Data-frame by Rows}
\usage{
rbinds(F1, F2, fill = 0)
}
\arguments{
\item{F1}{data.frame.}

\item{F2}{data.frame.}

\item{fill}{character/numeric. Default 0}
}
\value{
a combined data.frame
}
\description{
Combine two data.frame by rows by filling in missing column from each other based on \code{colnames}.
}
\examples{
F1 = data.frame(A = 1:3, B = 1:3)
F2 = data.frame(B = 1:3, C = 1:3)
F3 = data.frame(C = 1:3, D = 1:3)
Reduce(rbinds, list(F1, F2, F3))

}
