% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.r
\name{Read10xs}
\alias{Read10xs}
\title{Read10x in groups}
\usage{
Read10xs(
  airr_files = NULL,
  contig_files = NULL,
  consensus_files = NULL,
  clonotype_files = NULL,
  names = NULL,
  group = NULL,
  clonotype_colnames = NULL,
  consensus_colnames = NULL,
  unique_clonotype = TRUE,
  unique_clonotype_by_cdr3nt = TRUE,
  cdr3nt_colnames = NULL,
  cdr3aa_colnames = NULL,
  verbose = TRUE
)
}
\arguments{
\item{airr_files}{character. Path to AIRR files.}

\item{contig_files}{character. Path to contig_annotations files generated by 10x cellranger.}

\item{consensus_files}{character. Path to consensus_annotations files generated by 10x cellranger.}

\item{clonotype_files}{character. Path to clonotypes files generated by 10x cellranger.}

\item{names}{character. Sample names. Default number index.}

\item{group}{list. A list named by group names and contain corresponding sample names.}

\item{clonotype_colnames}{character. Colnames of clonotype_id. Default \code{clonotype_id}}

\item{consensus_colnames}{character. Colnames of consensus_id. Default \code{consensus_id}}

\item{unique_clonotype}{logical. Make clonotypes unique in group. Default \code{TRUE}}

\item{unique_clonotype_by_cdr3nt}{logical. Unique clonotypes by same cdr3 DNA, if FALSE, by same cdr3 Amino-acid. Default \code{TRUE}}

\item{cdr3nt_colnames}{character. Colnames of cdr3_nt. Default \code{cdr3_nt}}

\item{cdr3aa_colnames}{character. Colnames of cdr3_aa. Default \code{cdr3}}

\item{verbose}{logical. Print progress. Default \code{TRUE}.}
}
\value{
a Trust raw data object (TrustRaw object)
}
\description{
Read10x in groups
}
\examples{
print('waiting...')

}
