% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{RenameVdjSample}
\alias{RenameVdjSample}
\title{Reame an object of VDJ sample}
\usage{
RenameVdjSample(sample, name = NULL, verbose = TRUE)
}
\arguments{
\item{sample}{VDJsample. an object of VDJ sample}

\item{name}{character. a name}

\item{verbose}{logical.}
}
\value{
A named object of VDJ sample
}
\description{
Reame an object of VDJ sample
}
\examples{

}
