% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Sankey}
\alias{Sankey}
\title{Plot VJab Sankey}
\usage{
Sankey(VJab, out = NULL)
}
\arguments{
\item{VJab}{character}

\item{out}{character}
}
\value{

}
\description{
Plot VJab Sankey
}
\examples{
VJab = c('TRAV13-1~TRAJ49~TRBV20-1~TRBJ2-2', 'TRAV14DV4~TRAJ13~TRBV20-1~TRBJ2-5', 
         'TRAV1-2~TRAJ33~TRBV6-4~TRBJ2-3',   'TRAV5~TRAJ3~TRBV20-1~TRBJ2-3' )
Sankey(VJab)

}
