% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{sepInteger}
\alias{sepInteger}
\title{Separate Positive Integer}
\usage{
sepInteger(x, sep = NULL)
}
\arguments{
\item{x}{numeric.}

\item{sep}{integer.}
}
\value{
classification
}
\description{
Separate Positive Integer
}
\examples{
sepInteger(1:20, c(1,2,3))

}
