% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,clonotype-method}
\alias{show,clonotype-method}
\title{Overview of the Clonotype Class}
\usage{
\S4method{show}{clonotype}(object)
}
\arguments{
\item{object}{class.}

\item{clonotype}{class. An object of the clonotype class}
}
\value{
Brief information about a consensus object
}
\description{
Overview of the Clonotype Class
}
