% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,consensus-method}
\alias{show,consensus-method}
\title{Overview of the consensus class}
\usage{
\S4method{show}{consensus}(object)
}
\arguments{
\item{object}{class.}

\item{consensus}{class. An object of the consensus class}
}
\value{
Brief information about an consensus object
}
\description{
Overview of the consensus class
}
