% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,Trust-method}
\alias{show,Trust-method}
\title{Overview of the Trust class}
\usage{
\S4method{show}{Trust}(object)
}
\arguments{
\item{object}{class.}

\item{Trust}{class. An object of the Trust class}
}
\value{
Brief information about an consensus object
}
\description{
Overview of the Trust class
}
