% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,TrustRaw-method}
\alias{show,TrustRaw-method}
\title{Overview of the TrustRaw class}
\usage{
\S4method{show}{TrustRaw}(object)
}
\arguments{
\item{object}{class.}

\item{TrustRaw}{class. An object of the TrustRaw class}
}
\value{
Brief information about a TrustRaw object
}
\description{
Overview of the TrustRaw class
}
