% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,VDJ-method}
\alias{show,VDJ-method}
\title{Overview of the VDJ Class}
\usage{
\S4method{show}{VDJ}(object)
}
\arguments{
\item{object}{class.}

\item{VDJ}{class. An object of the VDJ class}
}
\value{
Brief information about a VDJ object
}
\description{
Overview of the VDJ Class
}
