% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{show,VDJsample-method}
\alias{show,VDJsample-method}
\title{Overview of the VDJsample Class}
\usage{
\S4method{show}{VDJsample}(object)
}
\arguments{
\item{object}{class.}

\item{VDJsample}{class. An object of the VDJsample class}
}
\value{
Brief information about a VDJsample object
}
\description{
Overview of the VDJsample Class
}
