% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{subsetCloneSize}
\alias{subsetCloneSize}
\title{Subset clonotype by clone size}
\usage{
subsetCloneSize(vdj, names = NULL, clone.size = c(1, Inf))
}
\arguments{
\item{vdj}{an object of VDJ.}

\item{names}{character. sample or group names.}

\item{clone.size}{vector. c(min, max)}
}
\value{
a subset clonotype VDJ object.
}
\description{
Subset clonotype by clone size
}
\examples{
VDJ = readRDS('VDJ.rds')
subsetCloneSize(vdj, clone.size = c(2, Inf))

}
