% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{subsetClonotype}
\alias{subsetClonotype}
\title{Subset Clonotype Object}
\usage{
subsetClonotype(clonotype, i)
}
\arguments{
\item{clonotype}{class.  an object of clonotype class}

\item{i}{vector. logical expression indicating position to keep}
}
\value{
a subset clonotype object
}
\description{
Subset Clonotype Object
}
\examples{

}
