% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{subsetConsensus}
\alias{subsetConsensus}
\title{Subset Consensus Object}
\usage{
subsetConsensus(consensus, i)
}
\arguments{
\item{consensus}{class.  an object of the consensus class}

\item{i}{vector. logical expression indicating position to keep}
}
\value{
a subset consensus object
}
\description{
Subset Consensus Object
}
\examples{

}
