% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableCDR3}
\alias{TableCDR3}
\title{Table CDR3dna or CDR3aa in Each Name}
\usage{
TableCDR3(
  vdj,
  names = NULL,
  type = NULL,
  target = NULL,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{object.}

\item{names}{character.}

\item{type}{character.}

\item{target}{character.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Table CDR3dna or CDR3aa in Each Name
}
\examples{
data = TableCDR3(vdj)
head(data)

}
