% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableVDJ}
\alias{TableVDJ}
\title{Table VDJ Genes in Each Name}
\usage{
TableVDJ(
  vdj,
  target = NULL,
  type = NULL,
  names = NULL,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{object.}

\item{target}{character.}

\item{type}{character.}

\item{names}{character.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Table VDJ Genes in Each Name
}
\examples{
data = TableVDJ(VDJ, vdj)
head(data)

}
