% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableVJab}
\alias{TableVJab}
\title{Table VJ-AB in Each Name}
\usage{
TableVJab(vdj, target = NULL, names = NULL, save = TRUE, out.pref = NULL)
}
\arguments{
\item{vdj}{object.}

\item{target}{character.}

\item{names}{character.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Table VJ-AB in Each Name
}
\examples{
data = TableVJab(vdj)
head(data)

}
