% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableVJpair}
\alias{TableVJpair}
\title{Table VJ Pairs in Each Name}
\usage{
TableVJpair(vdj, names = NULL, target = NULL, save = TRUE, out.pref = NULL)
}
\arguments{
\item{vdj}{object.}

\item{names}{character.}

\item{target}{character.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Table VJ Pairs in Each Name
}
\examples{
data = TableVJpair(vdj)
head(data)

}
