% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{topClone}
\alias{topClone}
\title{Show Top Clonotypes}
\usage{
topClone(
  vdj,
  names = NULL,
  n.top = 5,
  plot = TRUE,
  repel = TRUE,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{an object of vdj.}

\item{names}{character.}

\item{n.top}{numeric}

\item{plot}{logical.}

\item{repel}{logical.}

\item{save}{logical.}

\item{out.pref}{character.}
}
\value{

}
\description{
Show Top Clonotypes
}
\examples{

}
