% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneSize.r
\name{topClonotypes}
\alias{topClonotypes}
\title{Show Top Clonotypes}
\usage{
topClonotypes(
  vdj,
  names = NULL,
  n.top = NULL,
  plot = TRUE,
  repel = TRUE,
  save = TRUE
)
}
\arguments{
\item{vdj}{an object of vdj.}

\item{names}{character.}

\item{n.top}{numeric}

\item{plot}{logical.}

\item{repel}{logical.}

\item{save}{logical.}
}
\value{

}
\description{
Show Top Clonotypes
}
\examples{

}
