% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{Trust-class}
\alias{Trust-class}
\alias{Trust}
\title{The Trust class}
\value{
An object of the trust4 class
}
\description{
The Trust object is the center of each single-cell immune repertoire analysis.
slots are listed below:
}
\section{Slots}{

\describe{
\item{\code{barcode}}{character. Cell barcode in single-cell sequencing, eg: Sample1_ATGCCAGAACGACT.}

\item{\code{celltype}}{character. Inferred cell type, such as: abT, gdT or B.}

\item{\code{Achain}}{consensus. confident TCR/BCR Alpha-chain object.}

\item{\code{Bchain}}{consensus. confident TCR/BCR Beta-chain object.}

\item{\code{Achain2}}{list.      secondary TCR/BCR Alpha-chain objects.}

\item{\code{Bchain2}}{list.      secondary TCR/BCR Beta-chain objects.}
}}

