% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{TrustRaw-class}
\alias{TrustRaw-class}
\alias{TrustRaw}
\title{The Trust raw data class}
\value{
An object of the TrustRaw class
}
\description{
The TrustRaw object is used for raw data storage.
Slots are listed below:
}
\section{Slots}{

\describe{
\item{\code{samples}}{list. Raw data for each sample}

\item{\code{groups}}{list. Raw data for each group}

\item{\code{info}}{list. Raw data meta information}
}}

