% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{UniqueClonotype}
\alias{UniqueClonotype}
\title{Unique Clonotype Object}
\usage{
UniqueClonotype(clonotype)
}
\arguments{
\item{clonotype}{class.}
}
\value{
a unique clonotype object
}
\description{
Unique Clonotype Object
}
\examples{

}
