% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{Upset}
\alias{Upset}
\title{Plot UpSet}
\usage{
Upset(x, title = NULL, out = NULL, w = 10, h = 6)
}
\arguments{
\item{x}{list. a named list}

\item{title}{character.}

\item{out}{character.}

\item{w}{numeric.}

\item{h}{numeric.}
}
\value{

}
\description{
Plot UpSet
}
\examples{
Upset(list(A = 1:5, B = 2:6, C = 3:10))


}
