% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{VDJ-class}
\alias{VDJ-class}
\alias{VDJ}
\title{The VDJ class}
\value{
An object of the VDJ class
}
\description{
The VDJ class
}
\section{Slots}{

\describe{
\item{\code{samples}}{list. Raw data for each sample}

\item{\code{groups}}{list. Raw data for each group}

\item{\code{info}}{list. Raw data meta information}
}}

