% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{VDJsample-class}
\alias{VDJsample-class}
\alias{VDJsample}
\title{The VDJ Sample Class}
\value{
An object of the VDJsample class
}
\description{
The VDJsample object is used for sample consensus and clonotype storage.
Slots are listed below:
}
\section{Slots}{

\describe{
\item{\code{consensus}}{consensus object.}

\item{\code{clonotype}}{clonotype object.}

\item{\code{name}}{character.}
}}

