% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refIMGT.r
\name{build_IMGT_reference}
\alias{build_IMGT_reference}
\title{Build IMGT database reference}
\usage{
build_IMGT_reference(outdir = NULL, verbose = TRUE)
}
\arguments{
\item{outdir}{character. Default \code{getwd()}}

\item{verbose}{logical. Default TRUE}
}
\value{
if success, return TRUE
}
\description{
Build IMGT database reference
}
\examples{
\donttest{build_IMGT_reference('IMGT_reference')}

}
