% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{cbinds}
\alias{cbinds}
\title{Combine Two Data-frame by Columns}
\usage{
cbinds(F1, F2, fill = 0)
}
\arguments{
\item{F1}{data.frame.}

\item{F2}{data.frame.}

\item{fill}{character/numeric. Default 0}
}
\value{
a combined data.frame
}
\description{
Combine two data.frame by columns by filling in missing rows from each other based on \code{rownames}.
}
\examples{
F1 = data.frame(A = 1:10, B = 1:10, row.names = 1:10)
F2 = data.frame(C = 1:5,  D = 1:5,  row.names = 3:7)
F3 = data.frame(E = 1:7,  F = 1:7,  row.names = 5:11)
Reduce(cbinds, list(F1, F2, F3))

}
