% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotypeFunc.r
\name{CloneSize}
\alias{CloneSize}
\title{Estimate Clonotype clone size}
\usage{
CloneSize(vdj, names = NULL, sep = NULL, plot = TRUE, save = TRUE)
}
\arguments{
\item{vdj}{an object of VDJ.}

\item{names}{character. sample or group names.}

\item{sep}{integer.}

\item{plot}{logical.}

\item{save}{logical.}
}
\value{
results of clone size
}
\description{
Estimate Clonotype clone size
}
\examples{
VDJ = readRDS('VDJ.rds')
CloneSize(VDJ)

}
