% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{consensus-class}
\alias{consensus-class}
\alias{consensus}
\title{The Consensus Class}
\value{
an object of the consensus class
}
\description{
The Consensus Class
}
\section{Slots}{

\describe{
\item{\code{ID}}{character.}

\item{\code{ClonoID}}{character.}

\item{\code{Vgene}}{character.}

\item{\code{Dgene}}{character.}

\item{\code{Jgene}}{character.}

\item{\code{Cgene}}{character.}

\item{\code{Chain}}{character.}

\item{\code{FWR1dna}}{character.}

\item{\code{FWR1aa}}{character.}

\item{\code{CDR1dna}}{character.}

\item{\code{CDR1aa}}{character.}

\item{\code{FWR2dna}}{character.}

\item{\code{FWR2aa}}{character.}

\item{\code{CDR2dna}}{character.}

\item{\code{CDR2aa}}{character.}

\item{\code{FWR3dna}}{character.}

\item{\code{FWR3aa}}{character.}

\item{\code{CDR3dna}}{character.}

\item{\code{CDR3aa}}{character.}

\item{\code{FWR4dna}}{character.}

\item{\code{FWR4aa}}{character.}

\item{\code{iUMI}}{character.}

\item{\code{UMIs}}{numeric.}

\item{\code{iRead}}{character.}

\item{\code{Reads}}{numeric.}

\item{\code{Cells}}{numeric.}

\item{\code{Barcodes}}{character.}

\item{\code{FullLength}}{logical.}

\item{\code{CDR3germlineSimilarity}}{numeric.}
}}

