% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.r
\name{distribute}
\alias{distribute}
\title{distribution of round numeric}
\usage{
distribute(x, digits = NULL, plot = FALSE)
}
\arguments{
\item{x}{numeric.}

\item{digits}{numeric.}

\item{plot}{logical. Default \code{FALSE}}
}
\value{
a data.frame of round frequency.
}
\description{
distribution of round numeric
}
\examples{
distribute(sample(1:5,20, replace = TRUE))

}
