% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{fetchVJab}
\alias{fetchVJab}
\title{Fetch VJ-AB in a Consensus Class}
\usage{
fetchVJab(consensus, clonotype, verbose = TRUE)
}
\arguments{
\item{consensus}{class.   an object of the consensus class}

\item{clonotype}{class.   an object of the clonotype class}

\item{verbose}{logical.}
}
\value{

}
\description{
Fetch VJ-AB in a Consensus Class
}
\examples{
vjAB = fetchVJab(consensus, clonotype)
head(vjAB)

}
