% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{findCDR3Motif}
\alias{findCDR3Motif}
\title{Find CDR3 Motif in K-mer and Position}
\usage{
findCDR3Motif(cdr3, kmers = NULL, position = NULL)
}
\arguments{
\item{cdr3}{character.}

\item{kmers}{numeric.}

\item{position}{numeric.}
}
\value{

}
\description{
Find CDR3 Motif in K-mer and Position
}
\examples{
CDR3  = TableCDR3(vdj)
CDR3  = rep(CDR3$CDR3, CDR3$TotalCell)
Motif = findCDR3Motif(CDR3, 3, c(4, Inf))
head(Motif) 

}
