% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.r
\name{getCDS}
\alias{getCDS}
\title{Get CDS Sequence from a Fasta file by GTF file}
\usage{
getCDS(fa, gtf, out = "cds.fa", verbose = TRUE)
}
\arguments{
\item{fa}{character.}

\item{gtf}{character.}

\item{out}{character.}

\item{verbose}{logical.}
}
\value{

}
\description{
Get CDS Sequence from a Fasta file by GTF file
}
\examples{
print('waiting...')

}
