% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.r
\name{getCDS}
\alias{getCDS}
\title{Get CDS Sequence from a Fasta file by GTF file}
\usage{
getCDS(
  fa,
  gtf,
  cds.out = "cds.fa",
  pep.out = "pep.fa",
  init.codon = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{fa}{character. fasta(.gz) file path}

\item{gtf}{character. gtf(.gz) file path}

\item{cds.out}{character.}

\item{pep.out}{character.}

\item{init.codon}{logical.}

\item{verbose}{logical.}
}
\value{

}
\description{
Get CDS Sequence from a Fasta file by GTF file
}
\examples{
print('waiting...')

}
