% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.r
\name{htmlHref}
\alias{htmlHref}
\title{Get HTML href content}
\usage{
htmlHref(file, css = "body a")
}
\arguments{
\item{file}{character. Path to HTML file.}

\item{css}{character. HTML elements (one of css) to select. Default \code{'body a'}}
}
\value{
A character vector.
}
\description{
Get HTML href content
}
\examples{
\donttest{
Download('https://ftp.ensembl.org/pub', 'test.html')
content = htmlHref('test.html')
file.remove('test.html')
head(content)
}

}
