% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{Ifnull}
\alias{Ifnull}
\alias{\%||\%}
\title{Default for NULL value}
\usage{
x \%||\% y
}
\arguments{
\item{x}{ANY. An object}

\item{y}{ANY. A default value}
}
\value{
\code{\%||\%}: \code{x} unless \code{NULL}, otherwise \code{y}
}
\description{
set default value for object, equal to \code{\%||\%} in rlang package
}
\examples{
1    \%||\% 1
NA   \%||\% 1
NULL \%||\% 1

}
