% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{makePct}
\alias{makePct}
\title{Show as a percentage}
\usage{
makePct(x, digits = 3)
}
\arguments{
\item{x}{vector}

\item{digits}{numeric}
}
\value{

}
\description{
Show as a percentage
}
\examples{
makePct(.253)

}
