% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.r
\name{MergeVdjSamples}
\alias{MergeVdjSamples}
\title{Merge VDJ Sample Objects}
\usage{
MergeVdjSamples(samples, name = NULL, unique_clonotype = TRUE, verbose = TRUE)
}
\arguments{
\item{samples}{list.}

\item{name}{character.}

\item{unique_clonotype}{logical.}

\item{verbose}{logical.}
}
\value{
An merged object of VDJ sample.
}
\description{
Merge VDJ Sample Objects
}
\examples{

}
