% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{min_max}
\alias{min_max}
\title{Dispersion Normalization}
\usage{
min_max(x)
}
\arguments{
\item{x}{numeric.}
}
\value{
data in 0-1 after linear transformation
}
\description{
Dispersion Normalization
}
\examples{
min_max(1:5)

}
