% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10x.r
\name{Read10xs}
\alias{Read10xs}
\title{Read in batches via read10x}
\usage{
Read10xs(
  airr_files = NULL,
  contig_files = NULL,
  consensus_files = NULL,
  clonotype_files = NULL,
  names = NULL,
  group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{airr_files}{}

\item{contig_files}{}

\item{consensus_files}{}

\item{clonotype_files}{}

\item{names}{}

\item{group}{}

\item{verbose}{}
}
\value{

}
\description{
waiting for process
}
\examples{

}
