% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableVDJ}
\alias{TableVDJ}
\title{Table VDJ Genes in Each Name}
\usage{
TableVDJ(
  vdj,
  target = NULL,
  type = NULL,
  names = NULL,
  save = TRUE,
  out.pref = NULL
)
}
\arguments{
\item{vdj}{}

\item{target}{}

\item{type}{}

\item{names}{}

\item{save}{}

\item{out.pref}{}
}
\value{

}
\description{
Table VDJ Genes in Each Name
}
\examples{

}
