% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{TableVJab}
\alias{TableVJab}
\title{Table VJ-AB in Each Name}
\usage{
TableVJab(
  vdj,
  names = NULL,
  target = NULL,
  save = TRUE,
  out.pref = NULL,
  verbose = TRUE
)
}
\arguments{
\item{vdj}{object.}

\item{names}{character.}

\item{target}{character.}

\item{save}{logical.}

\item{out.pref}{character.}

\item{verbose}{logical.}
}
\value{

}
\description{
Table VJ-AB in Each Name
}
\examples{
data = TableVJab(vdj)
head(data)

}
